refactor file organization and standard naming
