+++ backend/entities/academic_advising/__init__.py
@@ -1,2 +1,4 @@
 from .drop_in_entity import DropInEntity
 from .drop_in_reminder_entity import DropInReminderEntity
+from .document_entity import DocumentEntity
+from .document_section_entity import DocumentSectionEntity
