+++ backend/models/academic_advising/__init__.py
@@ -1,2 +1,4 @@
 from .drop_in import DropIn
 from .drop_in_reminder import DropInReminder
+from .document_details import DocumentDetails
+from .document_section import DocumentSection
