+++ backend/models/academic_advising/document.py
@@ -1,3 +1,5 @@
+from pydantic import BaseModel
+
 """Pydantic models for the `Document` entity."""
 
 
@@ -5,7 +7,6 @@ __author__ = ["Nathan Kelete"]
 __copyright__ = "Copyright 2023"
 __license__ = "MIT"
 
-from pydantic import BaseModel
 
 class Document(BaseModel):
     id: int
